Modularize database runtime concerns
Extract E2EE CloudSync hooks and database opening into focused runtime modules without changing transport behavior.
